requirements.txt
# AI Resume & Portfolio Builder - Dependencies

# Core AI/ML
google-generativeai==0.8.3

# Document Processing
python-docx==1.1.2
PyPDF2==3.0.1

# Environment & Configuration
python-dotenv==1.0.1

# Web Interface
streamlit==1.39.0

# Data Processing (if needed)
# pandas==2.2.0
# numpy==1.26.4
